feat(app-page-builder): introduce editor state initializer
